refactor(packages/explorer/src/views/Mining/index.js packages/explorer/src/views/Mining/instructions
Move components to their own files
