+++ packages/explorer/src/views/Mining/image.js
@@ -1,0 +1,14 @@
+import * as React from 'react'
+/**
+ * Exports a class for displaying an image
+ */
+class Image extends React.Component {
+  constructor(props) {
+    super(props)
+  }
+  render() {
+    return <img src={this.props.imgSrc} alt={this.props.altText} />
+  }
+}
+
+export { Image }
